ethers: Replace example AccountKeyLegacy_01
