+++ ethers-ext/package.json
@@ -21,6 +21,7 @@
     "ethers": "^5.7.2"
   },
   "devDependencies": {
+    "@klaytn/ethers-ext": "^0.9.5-beta",
     "@types/aes-js": "^3.1.1",
     "@types/bn.js": "^5.1.1",
     "@types/chai": "^4.3.4",
